reorganized classes into packages
