+++ core/src/main/java/io/github/CR/PlagueRats/GUI_thaddeus/components/GameButton.java
@@ -1,4 +1,4 @@
-package io.github.CR.PlagueRats.GUI_thaddeus;
+package io.github.CR.PlagueRats.GUI_thaddeus.components;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
