+++ core/src/main/java/io/github/CR/PlagueRats/GUI_thaddeus/components/MenuManager.java
@@ -1,8 +1,6 @@
-package io.github.CR.PlagueRats.GUI_thaddeus.input;
+package io.github.CR.PlagueRats.GUI_thaddeus.components;
 
 // Handles UIWindow instances, auto-hides on show
-
-import io.github.CR.PlagueRats.GUI_thaddeus.UIWindow;
 
 import java.util.ArrayList;
 import java.util.List;
