+++ core/src/main/java/io/github/CR/PlagueRats/GUI_thaddeus/control/GameController.java
@@ -1,4 +1,4 @@
-package io.github.CR.PlagueRats.GUI_thaddeus;
+package io.github.CR.PlagueRats.GUI_thaddeus.control;
