+++ core/src/main/java/io/github/CR/PlagueRats/GUI_thaddeus/control/GameStage.java
@@ -1,4 +1,4 @@
-package io.github.CR.PlagueRats.GUI_thaddeus;
+package io.github.CR.PlagueRats.GUI_thaddeus.control;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.scenes.scene2d.Actor;
@@ -6,7 +6,11 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 
-import io.github.CR.PlagueRats.GUI_thaddeus.input.MenuManager;
+import io.github.CR.PlagueRats.GUI_thaddeus.input.CameraWrapper;
+import io.github.CR.PlagueRats.GUI_thaddeus.components.MenuManager;
+import io.github.CR.PlagueRats.GUI_thaddeus.components.StatsPanel;
+import io.github.CR.PlagueRats.GUI_thaddeus.render.CharacterActor;
+import io.github.CR.PlagueRats.GUI_thaddeus.render.SpriteProvider;
 import io.github.CR.PlagueRats.backend.AbstractCharacter;
 import io.github.CR.PlagueRats.backend.Cell;
 import io.github.CR.PlagueRats.backend.Command;
@@ -15,7 +19,7 @@ import io.github.CR.PlagueRats.backend.MapGenerator;
 import java.util.List;
 
 public class GameStage extends Stage {
-    private final StatsPanel         stats;
+    private final StatsPanel stats;
     private final int cellSize;
     private final SpriteProvider spriteProvider;
