--- core/src/main/java/io/github/CR/PlagueRats/GUI_thaddeus/deleted/InteractionManager.java
@@ -1,18 +1,0 @@
-/*package io.github.CR.PlagueRats.GUI_thaddeus;
-
-// Centralizes routing of clicks
-
-import com.badlogic.gdx.math.Vector2;
-
-public class InteractionManager {
-    private final GameWorldInteractor gameWorldInteractor;
-
-    public InteractionManager(GameWorldInteractor gameWorldInteractor) {
-        this.gameWorldInteractor = gameWorldInteractor;
-    }
-
-    public void routeClick(Vector2 worldCoords) {
-        gameWorldInteractor.handleWorldClick(worldCoords);
-    }
-}
-*/
