--- core/src/main/java/io/github/CR/PlagueRats/GUI_thaddeus/deleted/WorldInputHandler.java
@@ -1,7 +1,0 @@
-/*package io.github.CR.PlagueRats.GUI_thaddeus;
-
-public interface WorldInputHandler {
-    boolean handleKey(int keycode);
-    boolean handleTouch(int x, int y, int button);
-}
-*/
